backend/requirements.txt
# ── Core Framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── Database ─────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0          # async SQLite driver

# ── Authentication & Security ────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Validation ───────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.0
email-validator==2.1.1

# ── AI Integration ───────────────────────────────────────────
google-generativeai==0.6.0  # Gemini API (free tier)

# ── HTTP Client ──────────────────────────────────────────────
httpx==0.27.0

# ── Task Scheduling (reminders) ──────────────────────────────
apscheduler==3.10.4

# ── Notifications ────────────────────────────────────────────
exponent-server-sdk==2.1.0  # Expo Push Notifications

# ── Utilities ────────────────────────────────────────────────
python-dotenv==1.0.1
pytz==2024.1
babel==2.15.0              # i18n / date-time localization
